run_watcher.sh: move `nix` shell out to `flake.nix`
This way `./run_watcher.sh` should be usable as is outside `nix` as long
as dependencies are there.

`nix-shell` was slow to start up (a lot of `.nix` code to evaluate) which
makes `.hs` file changes a bit slow to build and test. Moving the setup
phase to `nix develop` speeds build/run eval quite a bit.